requirements.txt
streamlit
PyPDF2
speechrecognition
langchain
python-dotenv
